+++ src/main/java/com/example/demo/controller/JobController.java
@@ -37,14 +37,14 @@ public class JobController {
 
     @GetMapping("/{id}")
     @LogActivity(value = "Get job")
-    public ResponseEntity<JobDto> getById(@PathVariable int id){
+    public ResponseEntity<JobDto> getById(@PathVariable Long id){
         JobDto jobDto = jobService.getById(id);
         if(jobDto != null) return ResponseEntity.ok(jobDto);
         return ResponseEntity.notFound().build();
     }
     @PutMapping("/{id}")
     @LogActivity(value = "Update job")
-    public ResponseEntity<JobDto> update(@Valid @PathVariable int id, @RequestBody JobDto jobDto){
+    public ResponseEntity<JobDto> update(@Valid @PathVariable Long id, @RequestBody JobDto jobDto){
         JobDto updatedJob = jobService.update(jobDto, id);
         if(updatedJob == null) {
             return ResponseEntity.notFound().build();
@@ -54,7 +54,7 @@ public class JobController {
 
     @DeleteMapping("/{id}")
     @LogActivity(value = "Delete job")
-    public ResponseEntity delete(@PathVariable int id){
+    public ResponseEntity delete(@PathVariable Long id){
         boolean deleted = jobService.delete(id);
         if (deleted) return ResponseEntity.noContent().build();
         return ResponseEntity.notFound().build();
